desktop/requirements.txt
paho-mqtt
scipy
matplotlib
pyqt5
